refaoctor(plugin-pug): use transorm API instead of chain
