+++ packages/plugin-pug/modern.config.ts
@@ -3,11 +3,5 @@ import { buildConfigWithMjs } from '../../scripts/modern.base.config';
 
 export default {
   plugins: [moduleTools()],
-  buildConfig: buildConfigWithMjs.map((config) => {
-    if (config.format === 'cjs') {
-      // add loader to entry
-      config.input = ['src/index.ts', 'src/loader.ts'];
-    }
-    return config;
-  }),
+  buildConfig: buildConfigWithMjs,
 };
