+++ packages/plugin-pug/src/index.ts
@@ -1,4 +1,3 @@
-import path from 'node:path';
 import { mergeChainedOptions } from '@rsbuild/shared';
 import type { RsbuildPlugin } from '@rsbuild/core';
 import type { Options as PugOptions } from 'pug';
@@ -14,19 +13,44 @@ export type PluginPugOptions = {
 export const pluginPug = (options: PluginPugOptions = {}): RsbuildPlugin => ({
   name: 'rsbuild:pug',
 
-  setup(api) {
-    api.modifyBundlerChain(async (chain, { CHAIN_ID }) => {
-      const pugOptions = mergeChainedOptions({
-        defaults: {},
-        options: options.pugOptions,
-      });
+  async setup(api) {
+    const VUE_SFC_REGEXP = /\.vue$/;
+    const { compile, compileClient } = await import('pug');
 
-      chain.module
-        .rule(CHAIN_ID.RULE.PUG)
-        .test(/\.pug$/)
-        .use(CHAIN_ID.USE.PUG)
-        .loader(path.resolve(__dirname, './loader'))
-        .options(pugOptions);
+    const pugOptions = mergeChainedOptions({
+      defaults: {
+        doctype: 'html',
+        compileDebug: false,
+      },
+      options: options.pugOptions,
     });
+
+    api.transform(
+      { test: /\.pug$/ },
+      ({ code, resourcePath, addDependency }) => {
+        const options = {
+          filename: resourcePath,
+          ...pugOptions,
+        };
+
+        // Compile pug to HTML for Vue compiler
+        if (VUE_SFC_REGEXP.test(resourcePath)) {
+          const template = compile(code, options);
+          const { dependencies } = template as unknown as {
+            dependencies: string[];
+          };
+
+          if (dependencies) {
+            dependencies.forEach(addDependency);
+          }
+
+          return template();
+        }
+
+        // Compile pug to JavaScript for html-webpack-plugin
+        const templateCode = compileClient(code, options);
+        return `${templateCode}; export default template;`;
+      },
+    );
   },
 });
