--- packages/plugin-pug/tests/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@rsbuild/tsconfig/base",
-  "compilerOptions": {
-    "baseUrl": "./",
-    "types": ["vitest/globals"]
-  },
-  "include": ["./**/*.ts"]
-}
